% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{covs.test}
\alias{covs.test}
\title{Covariate test function for the horseshoe domain}
\usage{
covs.test(x, y)
}
\arguments{
\item{x, y}{Points at which to evaluate the test function.}
}
\value{
Returns function evaluations.
}
\description{
Implements a finite area test function the horseshoe domain.
}
