% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mesh.1.5D}
\alias{plot.mesh.1.5D}
\title{Plot a mesh.1.5D object}
\usage{
\method{plot}{mesh.1.5D}(x, ...)
}
\arguments{
\item{x}{A \code{mesh.1.5D} object defining the triangular mesh, as generated by \code{create.mesh.1.5D}
or \code{refine.mesh.1.5D}.}

\item{...}{Arguments representing graphical options to be passed to \link[graphics]{par}.}
}
\description{
Plot a mesh.1.5D object, generated by \code{create.mesh.1.5D} or \code{refine.mesh.1.5D}.
}
