% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{quasicircle2D}
\alias{quasicircle2D}
\title{Quasicircle2D domain}
\description{
The boundary and interior nodes and connectivity matrix of a triangular mesh of a quasicircular domain, together 
with a non-stationary field observed over the nodes of the mesh.  
This dataset can be used to create a \code{mesh.2D} object with the function \code{create.mesh.2D} and to test
the smooth.FEM function.
}
