% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{quasicircle2Dareal}
\alias{quasicircle2Dareal}
\title{Quasicircle2Dareal domain}
\description{
The mesh of a quasicircular domain, together with a non-stationary field observed over seven circular subdomains and 
the incindence matrix defining the subdomains used by Azzimonti et. al 2015.   
This dataset can be used to test the smooth.FEM function for areal data.
}
\references{
Azzimonti, L., Sangalli, L. M., Secchi, P., Domanin, M., & Nobile, F. (2015). Blood flow velocity 
field estimation via spatial regression with PDE penalization. Journal of the American Statistical 
Association, 110(511), 1057-1071.
}
