% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{refine.by.splitting.mesh.1.5D}
\alias{refine.by.splitting.mesh.1.5D}
\title{Create a \code{mesh.1.5D} object by splitting each edge of a given mesh into two subedges.}
\usage{
refine.by.splitting.mesh.1.5D(mesh = NULL)
}
\arguments{
\item{mesh}{a \code{mesh.1.5D} object to split}
}
\value{
An object of class mesh.1.5D with splitted edges
}
\description{
Create a \code{mesh.1.5D} object by splitting each edge of a given mesh into two subedges.
}
