% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{refine.by.splitting.mesh.2D}
\alias{refine.by.splitting.mesh.2D}
\title{Create a \code{mesh.2D} object by splitting each triangle of a given mesh into four subtriangles.}
\usage{
refine.by.splitting.mesh.2D(mesh = NULL)
}
\arguments{
\item{mesh}{a \code{mesh.2D} object to split}
}
\value{
An object of class mesh.2D with splitted triangles
}
\description{
Create a \code{mesh.2D} object by splitting each triangle of a given mesh into four subtriangles.
}
