% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{refine.by.splitting.mesh.3D}
\alias{refine.by.splitting.mesh.3D}
\title{Create a \code{mesh.3D} object by splitting each tetrahedron of a given mesh into eight subtetrahedrons.}
\usage{
refine.by.splitting.mesh.3D(mesh = NULL)
}
\arguments{
\item{mesh}{a \code{mesh.3D} object to split}
}
\value{
An object of class mesh.3D with splitted tetrahedrons
}
\description{
Create a \code{mesh.3D} object by splitting each tetrahedron of a given mesh into eight subtetrahedrons.
}
