% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{refine.mesh.1.5D}
\alias{refine.mesh.1.5D}
\title{Refine 1D mesh}
\usage{
refine.mesh.1.5D(mesh, delta)
}
\arguments{
\item{mesh}{a \code{mesh.1.5D} object to refine}

\item{delta}{the maximum allowed length}
}
\value{
An object of class mesh.1.5D with refined edges
}
\description{
Refine 1D mesh
}
