% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sphere3Ddata}
\alias{sphere3Ddata}
\title{Sphere3Ddata}
\description{
A dataset with information about the connectivity matrix and the nodes locations of a sphere geometry. It containes:
\itemize{
	\item nodes. A #nodes-by-3 matrix specifying the locations of each node.
	\item tetrahedrons. A #tetrahedrons-by-4 matrix specifying the indices of the nodes in each tetrahedron.
   }
This dataset can be used to create a \code{MESH.3D} object with the function \code{create.MESH.3D}
}
