% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{split.mesh.2.5D}
\alias{split.mesh.2.5D}
\title{Create a \code{mesh.2.5D} object by splitting each triangle of a given mesh into four subtriangles.}
\usage{
\method{split}{mesh.2.5D}(mesh = NULL)
}
\arguments{
\item{mesh}{a \code{mesh.2.5D} object to split}
}
\value{
An object of class mesh.2.5D with splitted triangles
}
\description{
Create a \code{mesh.2.5D} object by splitting each triangle of a given mesh into four subtriangles.
}
