% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{create.mesh.3D}
\alias{create.mesh.3D}
\title{Create a \code{mesh.3D} object from the connectivty matrix and nodes locations}
\usage{
create.mesh.3D(nodes, tetrahedrons, order = 1)
}
\arguments{
\item{nodes}{A #nodes-by-3 matrix specifying the locations of each node.}

\item{tetrahedrons}{A #tetrahedrons-by-4 matrix specifying the indices of the nodes in each tetrahedrons.}

\item{order}{Order of the Finite Element basis. Only order = 1 is currently implemented.}
}
\value{
An object of the class \code{mesh.3D} with the following output:
\itemize{
\item{\code{nnodes}}{The #nodes in the mesh.}
\item{\code{ntetrahedrons}}{The #tetrahedrons in the mesh.}
\item{\code{nodes}}{A #nodes-by-3 matrix containing the x,y and z coordinate for each point of the mesh.}
\item{\code{tetrahedrons}}{A #tetrahedrons-by-4 matrix specifying the indices of the nodes in each tetrahedron of the mesh.}
\item{\code{order}}{It specifies the order of the associated Finite Element basis. When order = 1, each 
mesh tetrahedron is represented by 4 nodes (the tetrahedron vertices).}
}
}
\description{
Create a \code{mesh.3D} object from the connectivty matrix and nodes locations
}
\examples{
library(fdaPDE)

##Load the matrix nodes and tetrahedrons
data(sphere3Ddata)

nodes = sphere3Ddata$nodes
tetrahedrons = sphere3Ddata$tetrahedrons

##Create the triangulated mesh from the connectivity matrix and nodes locations
mesh=create.mesh.3D(nodes,tetrahedrons)
}
