% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{horseshoe2D}
\alias{horseshoe2D}
\title{Horseshoe domain}
\description{
The boundary and interior nodes and connectivity matrix of a triangular mesh of the horseshoe domain. 
This dataset can be used to create a \code{mesh.2D} object with the function \code{create.mesh.2D}.
}
